% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{add_blueprint}
\alias{add_blueprint}
\title{Add blueprint to a recipe}
\usage{
add_blueprint(base_recipe, blueprint)
}
\arguments{
\item{base_recipe}{A the basic recipe specified by the user. See (TODO reference).}

\item{blueprint}{A blueprint id. Use \code{\link{TODO function}} for a list of available blueprints.}
}
\description{
This function binds together the base definition of the recipe
specified by the user with a selected blueprint id.
}
\examples{

TODO examples

}
