% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explaination.R
\docType{data}
\name{Beta}
\alias{Beta}
\title{Beta Data}
\format{
A data frame with 1000 rows and 11 variables:
\describe{
\item{Final Grade (\%)}{Description of Final Grade (\%)}
\item{Café Location}{Description of Café Location}
\item{Study Time}{Description of Study Time}
\item{Year in School}{Description of Year in School}
\item{Hobbies}{Description of Hobbies}
\item{Primary Sport}{Description of Primary Sport}
\item{Living Arrangement}{Description of Living Arrangement}
\item{Part-time Work Hours}{Description of Part-time Work Hours}
\item{Study Group Participation}{Description of Study Group Participation}
\item{Hours of Sleep per Night}{Description of Hours of Sleep per Night}
\item{Has a Tutor}{Description of Has a Tutor}
}
}
\usage{
Beta
}
\description{
A dataset containing beta regression toy data.
}
\keyword{datasets}
