% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprint.R
\name{blueprint}
\alias{blueprint}
\title{Blueprint}
\usage{
blueprint(base_recipe, recipe_summary)
}
\arguments{
\item{base_recipe}{A the basic recipe specified by the user. See (TODO reference).}

\item{recipe_summary}{A summarized recipes object provided by \code{\link{summarise_recipe}}.}
}
\description{
This is an exemplary blueprint.
}
\examples{

TODO examples

}
