% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_worms_taxonomy.R
\name{collect_from_worms}
\alias{collect_from_worms}
\title{Add WoRMs taxonomy}
\usage{
collect_from_worms(species_names, fuzzy = FALSE)
}
\arguments{
\item{species_names}{Character vector with taxa names.}

\item{fuzzy}{Uses the TAXAMATCH algorithm to fuzzy match species names.}
}
\description{
This function collects WoRMS entries from their database using the reported taxa.
}
\details{
Searches in batches of 50.
}
\references{
worrms R-package
}
