% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cookbook-package.R
\docType{package}
\name{cookbook-package}
\alias{cookbook}
\alias{cookbook-package}
\title{cookbook: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Konrad Semsch \email{konrad.semsch@gmail.com}

}
\keyword{internal}
