% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_hex.R
\name{generate_hex}
\alias{generate_hex}
\title{Gera uma cor aleatória em formato hexadecimal ou RGB}
\usage{
generate_hex(format = "hex")
}
\arguments{
\item{format}{Define o formato de saída: \code{"hex"} (padrão) ou \code{"rgb"}.}
}
\value{
Uma string com o código hexadecimal ou uma matriz RGB.
}
\description{
Função para gerar uma cor aleatória em formato \strong{hexadecimal} (padrão) ou \strong{RGB}.
}
\examples{
generate_hex()         # retorna algo como "#A3B9F0"
generate_hex("rgb")    # retorna a matriz RGB correspondente

}
