% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_worms_taxonomy.R
\name{get_worms_taxonomy}
\alias{get_worms_taxonomy}
\title{Get WoRMs taxonomy}
\usage{
get_worms_taxonomy(species_names)
}
\arguments{
\item{species_names}{Character vector with taxa names.}
}
\description{
This function checks all reported taxa against the WoRMS databse.
}
\details{
This function collects WoRMS entries based on the given species names, and reports the valid name
and rank that is found. It first finds all direct matches, and then does a fuzzy match for queries that do
not have a direct match against the database.
}
\references{
worrms R-package
}
