% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Saúda um nome informado}
\usage{
hello(name)
}
\arguments{
\item{name}{Character. Nome da pessoa a ser saudada.}
}
\value{
Uma string com a saudação.
}
\description{
Função simples que retorna uma saudação no formato \code{"Hello <nome>"}.
}
\examples{
hello("Jenni")
hello("mundo")

}
