% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explaination.R
\docType{data}
\name{OLS}
\alias{OLS}
\title{OLS Data}
\format{
A data frame with 1000 rows and 11 variables:
\describe{
\item{Has_Job}{Description of Has_Job}
\item{Year_of_Study}{Description of Year_of_Study}
\item{Financially_Dependent}{Description of Financially_Dependent}
\item{Monthly_Allowance}{Description of Monthly_Allowance}
\item{Cooks_at_Home}{Description of Cooks_at_Home}
\item{Living_Situation}{Description of Living_Situation}
\item{Housing_Type}{Description of Housing_Type}
\item{Goes_Out_Spends_Money}{Description of Goes_Out_Spends_Money}
\item{Drinks_Alcohol}{Description of Drinks_Alcohol}
\item{Net_Money}{Description of Net_Money}
\item{Monthly_Earnings}{Description of Monthly_Earnings}
}
}
\usage{
OLS
}
\description{
A dataset containing ordinary least squares regression toy data.
}
\keyword{datasets}
