% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explaination.R
\docType{data}
\name{parametric_survival}
\alias{parametric_survival}
\title{Parametric Survival Data}
\format{
A data frame with 1000 rows and 6 variables:
\describe{
\item{id}{Unique subject identifier}
\item{age}{Age of the subject (years)}
\item{group}{Study group (A or B)}
\item{baseline_marker}{Baseline biomarker level (continuous)}
\item{comorbidity_count}{Number of comorbidities (integer)}
\item{time}{Observed survival time (event or censoring)}
\item{heart_attacks}{Event indicator (1 = event occurred, 0 = censored)}
}
}
\usage{
parametric_survival
}
\description{
A dataset containing parametric survival regression toy data.
}
\keyword{datasets}
