% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{select_blueprint}
\alias{select_blueprint}
\title{Select a given blueprint}
\usage{
select_blueprint(recipe_summary, blueprint)
}
\arguments{
\item{recipe_summary}{A summarized recipes object provided by \code{\link{summarise_recipe}}.}

\item{blueprint}{A blueprint id. Use \code{\link{TODO function}} for a list of available blueprints.}
}
\description{
This is a helper function that evaluates to a selected blueprint id
and binds it with the base recipe.
}
\examples{

TODO examples

}
