% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{summarise_recipe}
\alias{summarise_recipe}
\title{Summarise base recipe}
\usage{
summarise_recipe(base_recipe)
}
\arguments{
\item{base_recipe}{A the basic recipe specified by the user. See (TODO reference).}
}
\description{
This is a helper function extracts essential information from the base recipe
and passes it further into the selected bluprint(s).
}
\examples{

TODO examples

}
