% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_worms_taxonomy.R
\name{WORMS}
\alias{WORMS}
\title{Create and update local WORMS taxonomy database}
\usage{
WORMS(
  species_names,
  overwrite = FALSE,
  db_path = "~/Google Drive/Synced/Werk/Onderzoek/_Databases/WORMS",
  db_file = "WORMS.csv",
  meta_file = "WORMS_meta.csv"
)
}
\arguments{
\item{species_names}{Character vector with taxa names.}

\item{overwrite}{Set to TRUE if you want to force a completely new db.}

\item{db_path}{Path to local database.}

\item{db_file}{Name of local database file.}

\item{meta_file}{Name of local metadata file related to database.}
}
\description{
This function only collects WORMS entries from API if they are not yet included in the
local WORMS database.
}
\references{
worrms R-package
}
