% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explaination.R
\docType{data}
\name{Beta}
\alias{Beta}
\title{Beta Data}
\format{
A data frame with X rows and Y variables:
\describe{
\item{var1}{description of var1}
\item{var2}{description of var2}
}
}
\usage{
Beta
}
\description{
A dataset containing beta regression toy data.
}
\keyword{datasets}
